refactor(views): migrate artists to per-folder layout (commit 5/7)
Split src/views/artists.rs (1428 lines) into src/views/artists/{mod,update,view}.rs
following the established template:

- mod.rs: types (Page, Message, Action, ViewData), trait impls
  (HasCommonAction via macro, ViewPage explicit), Default + new(),
  tests for column visibility default + get/set round-trip
- update.rs: impl ArtistsPage { fn update } + the toggle-column-visibility
  test
- view.rs: impl ArtistsPage { fn view, fn render_artist_row,
  fn render_album_child_row } + the per-mode column helpers
  (artists_stars_visible / artists_plays_visible) + the four helper tests

Macro consumers:
- define_view_columns! for ArtistsColumn / Visibility (8 columns,
  all-on default to match the post-Plays-column layout).
- impl_has_common_action!(ArtistsAction { NavigateAndExpandAlbum }) —
  artists only navigates to Albums (single expand variant).

The ExpandCenter pre-check in update() that routes 2nd-tier album rows
to NavigateAndExpandAlbum is preserved verbatim — that's the cross-view
drill-down path, not behavioral churn.

cargo build / clippy --all-targets -D warnings / nightly fmt --check /
test --workspace all clean (662 + 368 + 1 doctest).
